move `parseDirective` out of `SourceCode`
