+++ lib/source-code/source-code.js
@@ -213,24 +213,6 @@ function isSpaceBetween(sourceCode, first, second, checkInsideOfJSXText) {
 //-----------------------------------------------------------------------------
 
 /**
- * Extract the directive and the justification from a given directive comment and trim them.
- * @param {string} value The comment text to extract.
- * @returns {{directivePart: string, justificationPart: string}} The extracted directive and justification.
- */
-function extractDirectiveComment(value) {
-    const match = /\s-{2,}\s/u.exec(value);
-
-    if (!match) {
-        return { directivePart: value.trim(), justificationPart: "" };
-    }
-
-    const directive = value.slice(0, match.index).trim();
-    const justification = value.slice(match.index + match[0].length).trim();
-
-    return { directivePart: directive, justificationPart: justification };
-}
-
-/**
  * Ensures that variables representing built-in properties of the Global Object,
  * and any globals declared by special block comments, are present in the global
  * scope.
@@ -921,7 +903,7 @@ class SourceCode extends TokenStore {
                 return false;
             }
 
-            const { directivePart } = extractDirectiveComment(comment.value);
+            const { directivePart } = commentParser.extractDirectiveComment(comment.value);
 
             const directiveMatch = directivesPattern.exec(directivePart);
 
@@ -962,15 +944,6 @@ class SourceCode extends TokenStore {
         varsCache.set("configGlobals", configGlobals);
     }
 
-    static parseDirective(commentValue) {
-        const { directivePart } = extractDirectiveComment(commentValue);
-        const match = directivesPattern.exec(directivePart);
-        const directiveText = match[1];
-        const directiveValue = directivePart.slice(match.index + directiveText.length);
-
-        return { directiveText, directiveValue };
-    }
-
     /**
      * Applies configuration found inside of the source code. This method is only
      * called when ESLint is running with inline configuration allowed.
@@ -986,7 +959,7 @@ class SourceCode extends TokenStore {
 
         this.getInlineConfigNodes().forEach(comment => {
 
-            const { directiveText, directiveValue } = SourceCode.parseDirective(comment.value);
+            const { directiveText, directiveValue } = commentParser.parseDirective(comment);
 
             switch (directiveText) {
                 case "exported":
